requirements_py312.txt
# Core PyTorch ecosystem
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face ecosystem
transformers>=4.35.0
diffusers>=0.24.0
huggingface-hub>=0.19.0
tokenizers>=0.15.0
accelerate>=0.24.0
safetensors>=0.4.0

# Core scientific computing
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
pillow>=10.0.0
scikit-image>=0.21.0

# Configuration and utilities
omegaconf>=2.3.0
tqdm>=4.66.0
einops>=0.7.0
pyyaml>=6.0.1

# Image processing and computer vision
opencv-python>=4.8.0
imageio>=2.31.0
albumentations>=1.3.0
kornia>=0.7.0
invisible-watermark>=0.2.0

# Deep learning utilities
pytorch-lightning>=2.1.0
torchmetrics>=1.2.0
torch-fidelity>=0.3.0

# Data handling
pandas>=2.1.0
pyarrow>=14.0.0

# Visualization and monitoring
tensorboard>=2.15.0
matplotlib-inline>=0.1.6

# Jupyter ecosystem
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipython>=8.15.0
jupyterlab>=4.0.0

# File handling and utilities
regex>=2023.10.0
ftfy>=6.1.0
filelock>=3.13.0
fsspec>=2023.10.0

# Optional dependencies for enhanced functionality
xformers>=0.0.22  # Memory efficient attention (optional)

# Development and debugging
ipdb>=0.13.13

# Compatibility
typing-extensions>=4.8.0

# Additional dependencies from original environment
requests>=2.31.0
certifi>=2023.7.22
urllib3>=2.0.0
